% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_pipeline.r
\name{annotate_data}
\alias{annotate_data}
\title{Annotates the data using given IDs}
\usage{
annotate_data(dat, id1, id2)
}
\arguments{
\item{dat}{Data.frame of data from vcf files or OpenGWAS DB}

\item{id1}{DatasetsID class of exposure IDs}

\item{id2}{DatasetsID class of outcome IDs}
}
\value{
Data.frame of annotated dat
}
\description{
Annotates the data using given IDs
}
