% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{annotate_efo}
\alias{annotate_efo}
\title{Annotate diseases with EFO IDs}
\usage{
annotate_efo(dat, column = "outcome")
}
\arguments{
\item{dat}{Data.frame of data}

\item{column}{Column name containing disease names}
}
\value{
Data.frame with appended column for EFO IDs
}
\description{
Attempts to annotate disease names with EFO IDs using the
`epigraphdb` package.
Note that the matching is fuzzy and some disease names will have multiple
associated EFOs which may differ in definition slightly.
}
