% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{annotate_ensg}
\alias{annotate_ensg}
\title{Annotate ENSG IDs with gene names}
\usage{
annotate_ensg(
  dat,
  column = "exposure",
  gene_name_col = "hgnc_symbol",
  build = "grch37"
)
}
\arguments{
\item{dat}{Data.frame of data}

\item{gene_name_col}{Column to append to `dat` with converted names (Optional)}

\item{build}{Genomic build (Optional)}

\item{col}{Column name containing ENSG IDs (Optional)}
}
\value{
Data.frame with appended column for names
}
\description{
Attempts to convert ENSG IDs to gene names (hgnc_symbol).
This is attempting using biomaRt's service and thus requires the optional
biomaRt package to be installed.
}
