% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{calc_f_stat}
\alias{calc_f_stat}
\title{Calculate F-statistic}
\usage{
calc_f_stat(dat, f_cutoff = 10, force_approx = FALSE, verbose = TRUE)
}
\arguments{
\item{dat}{Data.frame from do_mr()}

\item{f_cutoff}{F-statistic cutoff (Optional)}

\item{force_approx}{Force to use the approximate F-statistic instead (Optional, boolean)}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Modified `dat` data.frame (if f_cutoff > 0 supplied)
}
\description{
Calculates portion of variance explained and F-statistic.
If the data is lacking key information, i.e. allele frequencies, sample size
or consists of only one SNP, then the approximate F-statistic will be used
instead: \eqn{F = b ** 2 / SE ** 2}.
}
\seealso{
[do_mr()]
}
