% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{cis_trans}
\alias{cis_trans}
\title{Annotate cis/trans SNPs}
\usage{
cis_trans(
  dat,
  cis_region = 5e+05,
  chr_col = "chr.exposure",
  pos_col = "position.exposure",
  snp_col = NULL,
  values_col = "exposure",
  filter = "ensembl_gene_id",
  missing = "include",
  build = "grch37"
)
}
\arguments{
\item{dat}{Data.frame of data}

\item{cis_region}{Cis region definition (Optional, in kb)}

\item{chr_col}{Column name for chromosome (Optional)}

\item{pos_col}{Column name for position (Optional)}

\item{snp_col}{Column name for SNP rsIDs (Optional)}

\item{values_col}{Column name for gene names or ENSG IDs (Optional)
NB: Choice must match the `filter` value}

\item{filter}{How to search for genes in biomaRt, either:
\enumerate{
\item "ensembl_gene_id" for ENSG IDs, or
\item "hgnc_symbol" for gene names
}
(Optional)}

\item{missing}{"Include" or "drop" SNPs which could not be matched (Optional)}

\item{build}{Genomic build (Optional)}
}
\value{
Data.frame with appended column for cis/trans status
}
\description{
Attempts to annotate SNPs as cis or trans depending on their
location to the gene coding region.
This is achieved using the `biomaRt` R package.
}
