% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{clean_numeric}
\alias{clean_numeric}
\title{Remove * at the end of a vector}
\usage{
clean_numeric(x)
}
\arguments{
\item{x}{remove}
}
\value{
data.frame with query friendly names
}
\description{
Removes significance level inditcators (\code{*}), if present, and converts result to numeric.
}
\examples{
clean_numeric(raw_data <- c("0.003*","0.003**", "0.003"))
}
