% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{coloc_sub}
\alias{coloc_sub}
\title{Sub-function for the colocalisation analyses}
\usage{
coloc_sub(dat1, dat2, pairs, i, chrpos, report)
}
\arguments{
\item{dat1}{SNPs, etc. from first dataset}

\item{dat2}{SNPs, etc. from second dataset}

\item{pairs}{A data.frame of pairwise combined IDs}

\item{i}{Location in pairs being colocalised}

\item{chrpos}{Chromosome position to search for SNPs, default uses region
around lead SNP but can be overwritten if provided}

\item{report}{QCReport class of results, etc. for reporting}
}
\value{
Results, or empty if cannot run the colocalisation
}
\description{
Sub-function for the colocalisation analyses
}
