% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{combine_experiments_data}
\alias{combine_experiments_data}
\title{Combines multiple exeperiments data}
\usage{
combine_experiments_data(results)
}
\arguments{
\item{results}{\code{list} as returned from \code{\link{compute_mr}}}
}
\value{
data.frame with same structure as returned from \code{\link{combine_experiment_data}}
}
\description{
Takes a list of results for multiple traits, applies \code{\link{combine_experiment_data}} and combines the result
}
