% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{combine_results}
\alias{combine_results}
\title{Combine MR and coloc results into one data.frame}
\usage{
combine_results(
  mr_res,
  coloc_res,
  mr_res.by = c("id.exposure", "id.outcome"),
  coloc_res.by = c("file.exposure", "file.outcome")
)
}
\arguments{
\item{mr_res}{A data.frame of MR results from `do_mr()`}

\item{coloc_res}{A data.frame of coloc results from `do_coloc()`}

\item{mr_res.by}{MR columns to use for merging}

\item{coloc_res.by}{Coloc columns to use for merging}
}
\value{
Data.frame of merged results
}
\description{
Combine MR and coloc results into one data.frame
}
