% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_cochrane_pvalue}
\alias{compute_cochrane_pvalue}
\title{Cochrane pvalue}
\usage{
compute_cochrane_pvalue(q_ivw, nsnp)
}
\arguments{
\item{q_ivw}{\code{numeric}}

\item{nsnp}{\code{numeric}}
}
\value{
\code{numeric} pvalue
}
\description{
Cochrane pvalue
}
