% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_egger}
\alias{compute_egger}
\title{Compute Egger esitamtes}
\usage{
compute_egger(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
\code{data.frame} with \itemize{
  \item{effect}
  \item{se}
  \item{lb}
  \item{ub}
  \item{wald_test}
  \item{pval}
  \item{method }
  \item{or}
  \item{ci_low }
  \item{ci_high}
} columns
}
\description{
Compute Egger esitamtes
}
