% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_fstat}
\alias{compute_fstat}
\title{Compute F stat}
\usage{
compute_fstat(r2, n_exp)
}
\arguments{
\item{r2}{\code{numeric}}

\item{n_exp}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Compute F stat
}
