% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_i2}
\alias{compute_i2}
\title{Compute I2}
\usage{
compute_i2(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
\code{numeric}
}
\description{
Compute I2
}
