% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compute_meta}
\alias{compute_meta}
\title{Compute metadata}
\usage{
compute_meta(bxg, byg, seX, seY, mr, nsnp, n_out, n_cas, n_exp)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}

\item{mr}{\code{list} as returned from \code{\link{compute_mr}}}

\item{nsnp}{\code{numeric} number of values}

\item{n_out}{\code{numeric} as provided in \code{\link{run_mr}} metadata.}

\item{n_cas}{\code{numeric} as provided in \code{\link{run_mr}} metadata.}

\item{n_exp}{\code{numeric} as provided in \code{\link{run_mr}} metadata.}
}
\value{
a named \code{list} with four \code{numeric} fields
\itemize{
  \item{r2}
  \item{f.test}
  \item{power}
  \item{nsnp}
}
}
\description{
Compute metadata
}
