% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compute_mr}
\alias{compute_mr}
\title{Computes causal estimates}
\usage{
compute_mr(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
a named \code{list} with four \code{data.frame} fields
 \itemize{
 \item{ivw}
 \item{egger}
 \item{wm}
 \item{mbe}
}
Each data.frame contains following columns method, effect, or, ci_low, ci_high, pval
}
\description{
Computes causal estimates
}
