% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_power}
\alias{compute_power}
\title{Power of the study for a specific result i.e. OR}
\usage{
compute_power(r2, n_cas, n_out, ivw, alpha = 0.05)
}
\arguments{
\item{n_cas}{\code{numeric}}

\item{n_out}{\code{numeric}}

\item{ivw}{\code{numeric}}

\item{alpha}{\code{numeric}}

\item{rw}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Power of the study for a specific result i.e. OR
}
