% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_q_eg}
\alias{compute_q_eg}
\title{Rucker's Q test (p-value) of instrument i.e. trait}
\usage{
compute_q_eg(bxg, byg, seX, seY, nsnp, mreager)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}

\item{nsnp}{\code{numeric}}

\item{mreager}{\code{data.frame} as returned from \code{\link{compute_egger}}}
}
\value{
\code{numeric}
}
\description{
Rucker's Q test (p-value) of instrument i.e. trait
}
