% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_q_ivw}
\alias{compute_q_ivw}
\title{Compute q ivw}
\usage{
compute_q_ivw(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
numeric
}
\description{
Compute q ivw
}
