% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_r2}
\alias{compute_r2}
\title{Compute R^2}
\usage{
compute_r2(bxg, byg, seX, seY, n_out, n_cas, n_exp)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}

\item{n_out}{\code{numeric}}

\item{n_cas}{\code{numeric}}

\item{n_exp}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Compute R^2
}
