% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_ratios}
\alias{compute_ratios}
\title{Compute ratios}
\usage{
compute_ratios(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
a named \code{list} with two fields \itemize{
  \item{ratio} {bxg to byg ratio}
  \item{se_ratio} {seX to seY ratio}
}
}
\description{
Compute ratios
}
