% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compute_result_for_exposure}
\alias{compute_result_for_exposure}
\title{Compute MR results for a single exposure / trait}
\usage{
compute_result_for_exposure(df, exposure, experiment_meta)
}
\arguments{
\item{df}{data.frame A \code{data.frame} containing at least following columns
\itemize{
 \item{trait} {\code{character}}
 \item{effect_of_lead_variant_on_exposure_levels} {\code{numeric}}
 \item{effect_of_lead_variant_on_outcome_levels} {\code{numeric}}
 \item{standard_error_of_effect_on_exposure_se} {\code{numeric}}
 \item{standard_error_of_effect_on_outcome_se} {\code{numeric}}
}}

\item{exposure}{character. Should match one of the values in trait column of \code{df}}

\item{experiment_meta}{list A \code{list} with at least following
\itemize{
 \item{n_out} {\code{numeric}} of length equal to 1
 \item{n_cas} {\code{numeric}} of length equal to 1
 \item{n_exp} {A named \code{list}} of {\code{numeric}} values
}}
}
\value{
a named \code{list} with following items
\itemize{
 \item{mr} a named \code{list} as returned from \code{\link{compute_mr}}
 \item{heterogeneity_metrics} a named \code{list} as returned from \code{\link{compute_heterogeneity_metrics}} \code{heterogeneity_metrics}
 \item{heterogeneity_tests} a named \code{list} as returned from \code{\link{compute_heterogeneity_metrics}} \code{heterogeneity_tests}
 \item{meta} a named \code{list} as returned from \code{\link{compute_meta}}
 \item{data} a named \code{list} containg \itemize{
   \item{effect_of_lead_variant_on_exposure_levels} {\code{numeric}}
   \item{effect_of_lead_variant_on_outcome_levels} {\code{numeric}}
   \item{standard_error_of_effect_on_exposure_se} {\code{numeric}}
   \item{standard_error_of_effect_on_outcome_se} {\code{numeric}}
 }
}
}
\description{
Compute MR results for a single exposure / trait
}
