% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_ruckers_q_test_pvalue}
\alias{compute_ruckers_q_test_pvalue}
\title{Compute Rucker's Q test p-value}
\usage{
compute_ruckers_q_test_pvalue(q_eq, nsnp)
}
\arguments{
\item{q_eq}{\code{numeric}}

\item{nsnp}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Used by \code{\link{compute_heterogeneity_metrics}}
}
