% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_weighted_median}
\alias{compute_weighted_median}
\title{Compute weighted median}
\usage{
compute_weighted_median(bxg, byg, seX, seY)
}
\arguments{
\item{bxg}{\code{numeric} effect of lead variant on_exposure level}

\item{byg}{\code{numeric} effect of lead variant on outcome levels}

\item{seX}{\code{numeric} standard error of effect on exposure se}

\item{seY}{\code{numeric} standard error of effect on outcome se}
}
\value{
\code{data.frame} with \itemize{
  \item{effect}
  \item{or}
  \item{ci_low}
  \item{ci_high}
  \item{pval}
} columns
}
\description{
Compute weighted median
}
