% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{dat_to_gwasvcf}
\alias{dat_to_gwasvcf}
\title{Convert data.frame to gwasvcf format.}
\usage{
dat_to_gwasvcf(
  dat,
  out,
  chr_col,
  pos_col,
  nea_col,
  ea_col,
  snp_col = NULL,
  eaf_col = NULL,
  beta_col = NULL,
  se_col = NULL,
  pval_col = NULL,
  n = NULL,
  n_case = NULL,
  name = NULL,
  bcf_tools = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.frame}

\item{out}{Path to save output}

\item{chr_col}{Column name for chromosome}

\item{pos_col}{Column name for position}

\item{nea_col}{Column name for non-effect allele}

\item{ea_col}{Column name for effect allele}

\item{snp_col}{Column name for SNP (Optional)}

\item{eaf_col}{Column name for effect allele frequency (Optional)}

\item{beta_col}{Column name for beta (Optional)}

\item{se_col}{Column name for standard error (Optional)}

\item{pval_col}{Column name for P value (Optional)
NB: P values will be saved as 10^-P}

\item{n}{Sample size (Optional), can be int or column name}

\item{n_case}{Number of cases (Optional), can be int or column name}

\item{name}{Trait name (Optional), can be string or column name}

\item{bcf_tools}{Path to bcf_tools (Optional)}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
gwasvcf object
}
\description{
Function to convert a data.frame to gwasvcf format.
}
\seealso{
[file_to_gwasvcf()] for converting files.
}
