% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\docType{class}
\name{DatasetIDs-class}
\alias{DatasetIDs-class}
\alias{DatasetIDs}
\title{A Reference Class for the dataset IDs}
\description{
A Reference Class for the dataset IDs
}
\section{Fields}{

\describe{
\item{\code{info}}{A data.frame of details of each dataset, including filename,
annotated trait name and ID if from OpenGWAS DB}
}}

\section{Methods}{

\describe{
\item{\code{add_id(id, trait, filename)}}{Adds new ID with trait and filename}

\item{\code{annotate(id, name)}}{Annotates IDs, currently using biomaRt for ENSGs}

\item{\code{bad_id(id)}}{Notes that an ID is 'bad', i.e. no data loaded}

\item{\code{remove_bad_ids()}}{Removes those IDs from the data.frame which are 'bad'}
}}

