% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pqtl_decode.R
\name{decode_pqtl_file_name}
\alias{decode_pqtl_file_name}
\title{Title}
\usage{
decode_pqtl_file_name(unique_id, decode_linker_file, decode_dir)
}
\arguments{
\item{decode_dir}{}
}
\description{
Title
}
