% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{do_coloc}
\alias{do_coloc}
\title{Runs colocalisation using the coloc R package's coloc.abf function}
\usage{
do_coloc(dat, id1, id2, report, window = 400, chrpos = "", nthreads = 1)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{id1}{DatasetsID class of exposure IDs}

\item{id2}{DatasetsID class of outcome IDs}

\item{report}{QCReport class of results, etc. for reporting}

\item{window}{Region to extract SNPs from, either side of lead SNP,
default: 400kb}

\item{chrpos}{Chromosome position to search for SNPs, default uses region
around lead SNP but can be overwritten if provided}

\item{nthreads}{Number of threads for use in multithreaded functions,
default: 1}
}
\description{
Runs colocalisation using the coloc R package's coloc.abf function
}
