% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{do_mr}
\alias{do_mr}
\title{Runs Mendelian randomisation and related analyses, including heterogeneity,
Steiger filtering and pleiotropy analyses.
Also generates a number of plots, e.g. volcano plots, for MR results.}
\usage{
do_mr(dat, report)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{report}{QCReport class of results, etc. for reporting}
}
\value{
A data.frame of MR results
}
\description{
Runs Mendelian randomisation and related analyses, including heterogeneity,
Steiger filtering and pleiotropy analyses.
Also generates a number of plots, e.g. volcano plots, for MR results.
}
