% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.calc_f_stat}
\alias{.calc_f_stat}
\title{Calculates F-statistic and related}
\usage{
.calc_f_stat(dat, f_cutoff)
}
\arguments{
\item{dat}{A data.frame of data}
}
\value{
List of data.frame and integer
}
\description{
Calculates F-statistic and related
}
