% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.calc_pve}
\alias{.calc_pve}
\title{Calculate PVE}
\usage{
.calc_pve(b, maf, se, n)
}
\arguments{
\item{b}{Vector or number, beta}

\item{maf}{Vector or number, minor allele frequency}

\item{se}{Vector or number, standard error of beta}

\item{n}{Vector or number, sample size}
}
\value{
Vector or number, proportion of variance explained
}
\description{
Calculates proportion of variance explained.
From \href{https://doi.org/10.1371/journal.pone.0120758}{S1 Text}
}
\keyword{Internal}
