% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{.cdat_from_mixed}
\alias{.cdat_from_mixed}
\title{Helper function to extract colocalisation regions for when one dataset comes
from a local file and another from OpenGWAS.}
\usage{
.cdat_from_mixed(f1, f2, chrpos, verbose = TRUE)
}
\arguments{
\item{f1}{File path or OpenGWAS ID for trait 1}

\item{f2}{File path or OpenGWAS ID for trait 2}

\item{chrpos}{Character of the format chr:pos1-pos2}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
list of coloc-ready data
}
\description{
Helper function to extract colocalisation regions for when one dataset comes
from a local file and another from OpenGWAS.
}
\seealso{
[gwasglue::gwasvcf_to_coloc()], [gwasglue::ieugwasr_to_coloc()]
}
\keyword{Internal}
