% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_target.r
\name{.chunk}
\alias{.chunk}
\title{Splits vector into chunks}
\usage{
.chunk(x, n)
}
\arguments{
\item{x}{Vector}

\item{n}{Number of chunks to create}
}
\value{
list of chunks
}
\description{
Splits vector into chunks
}
\keyword{Internal}
