% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.coloc_sub}
\alias{.coloc_sub}
\title{Sub-function for the colocalisation analyses}
\usage{
.coloc_sub(
  dat1,
  dat2,
  min_snps = 100,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  susie = FALSE,
  bfile = NULL,
  plink = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat1}{SNPs, etc. from first dataset}

\item{dat2}{SNPs, etc. from second dataset}

\item{min_snps}{Number of minimum SNPs to check for analysis to continue (Optional)}

\item{p1}{p1 for coloc (Optional)}

\item{p2}{p2 for coloc (Optional)}

\item{p12}{p12 for coloc (Optional)}

\item{susie}{Run SuSiE? (Optional, boolean)}

\item{bfile}{Path to Plink bed/bim/fam files (Optional; required for SuSiE)}

\item{plink}{Path to Plink binary (Optional; required for SuSiE)}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Results data.frame
}
\description{
Sub-function for the colocalisation analyses
}
\keyword{Internal}
