% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.coloc_susie_sub}
\alias{.coloc_susie_sub}
\title{Sub function to run SuSiE and coloc}
\usage{
.coloc_susie_sub(d1, d2, bfile = NULL, plink = NULL, verbose = TRUE, ...)
}
\arguments{
\item{d1}{Dataset 1}

\item{d2}{Dataset 2}

\item{bfile}{Path to Plink bed/bim/fam files (Optional; required for SuSiE)}

\item{plink}{Path to Plink binary (Optional; required for SuSiE)}

\item{verbose}{Display verbose information (Optional, boolean)}

\item{...}{Other arguments passed to coloc.susie and coloc.bf}
}
\value{
Results data.frame
}
\description{
Sub function to run SuSiE and coloc
}
\keyword{Internal}
