% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{.combine_ids}
\alias{.combine_ids}
\title{Pairwise combination of two DatasetIDs classes}
\usage{
.combine_ids(id1, id2)
}
\arguments{
\item{id1}{DatasetsIDs class of exposure IDs}

\item{id2}{DatasetsIDs class of outcome IDs}
}
\value{
data.frame
}
\description{
Pairwise combination of two DatasetIDs classes
}
