% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_target.r
\name{.dgidb_linkage}
\alias{.dgidb_linkage}
\title{Link ENSG IDs with DGIdb}
\usage{
.dgidb_linkage(ensgs)
}
\arguments{
\item{ensgs}{Vector of ENSG IDs}
}
\value{
data.frame of results
}
\description{
Lookup ENSGs using the Drug Genome Interaction DB API.
}
\keyword{Internal}
