% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{.ensg_to_name}
\alias{.ensg_to_name}
\title{Converts ENSG to gene name using biomaRt}
\usage{
.ensg_to_name(ensg)
}
\arguments{
\item{ensg}{Vector of or single character, ENSG IDs}
}
\value{
Vector of or single character, gene names
}
\description{
Converts ENSG to gene name using biomaRt
}
