% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{.gwasvcf_to_coloc_rsid}
\alias{.gwasvcf_to_coloc_rsid}
\title{Prepare gwasvcf files for coloc.
This method will extract SNPs from one file using one chrompos and then look
up those SNPs in the other file -- this is to ensure coloc can be conducted
upon two datasets of different genomic builds without the need of liftover.}
\usage{
.gwasvcf_to_coloc_rsid(
  vcf1,
  vcf2,
  chrompos,
  type1 = NULL,
  type2 = NULL,
  build1 = "GRCh37",
  build2 = "GRCh37",
  verbose = TRUE
)
}
\arguments{
\item{vcf1}{VCF object or path to vcf file}

\item{vcf2}{VCF object or path to vcf file}

\item{chrompos}{Character of the format chr:pos1-pos2}
}
\value{
list of coloc-ready data, or NA if failed
}
\description{
Prepare gwasvcf files for coloc.
This method will extract SNPs from one file using one chrompos and then look
up those SNPs in the other file -- this is to ensure coloc can be conducted
upon two datasets of different genomic builds without the need of liftover.
}
\keyword{Internal}
