% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.gwasvcf_to_pwcoco}
\alias{.gwasvcf_to_pwcoco}
\title{Prepare gwasvcf files for PWCoCo}
\usage{
.gwasvcf_to_pwcoco(vcf1, vcf2, chrompos, type1 = NULL, type2 = NULL, outfile)
}
\arguments{
\item{vcf1}{VCF object or path to VCF file}

\item{vcf2}{VCF object or path to VCF file}

\item{chrompos}{Character of the format chr:pos1-pos2}

\item{type1}{How to treat vcffile1 for coloc, either "quant" or "cc" (Optional)}

\item{type2}{How to treat vcffile2 for coloc, either "quant" or "cc" (Optional)}

\item{outfile}{Path to output files, without file ending}
}
\value{
0 if success, 1 if there was a problem
}
\description{
Write files for PWCoCo where data are read from two VCF objects or files.
}
\keyword{Internal}
