% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.r
\name{.ivw_delta}
\alias{.ivw_delta}
\title{IVW weighted delta}
\usage{
.ivw_delta(dat)
}
\arguments{
\item{object}{Harmonised data.frame}
}
\value{
Results data.frame
}
\description{
Calculates the inverse variance weighted delta method from the MendelianRandomization package
}
\keyword{Internal}
