% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{.print_msg}
\alias{.print_msg}
\title{Helper function for message printing.}
\usage{
.print_msg(msg, verbose)
}
\arguments{
\item{msg}{Message}

\item{verbose}{Display message or suppress}
}
\description{
Helper function for message printing.
}
\keyword{Internal}
