% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.pwcoco_sub}
\alias{.pwcoco_sub}
\title{Sub-function to run PWCoCo}
\usage{
.pwcoco_sub(
  bfile,
  chrpos,
  pwcoco,
  maf = 0.01,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  workdir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{bfile}{Path to Plink bed/bim/fam files}

\item{chrpos}{Character of the format chr:pos1-pos2}

\item{pwcoco}{Path to PWCoCo executible}

\item{maf}{MAF cut-off (Optional)}

\item{p1}{p1 for coloc (Optional)}

\item{p2}{p2 for coloc (Optional)}

\item{p12}{p12 for coloc (Optional)}

\item{workdir}{Path to save temporary files (Optional)}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Results data.frame
}
\description{
Sub-function to run PWCoCo
}
\keyword{Internal}
