% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{.read_dataset}
\alias{.read_dataset}
\title{Read datasets}
\usage{
.read_dataset(
  ids,
  rsids = NULL,
  pval = 5e-08,
  proxies = TRUE,
  proxy_rsq = 0.8,
  proxy_kb = 5000,
  proxy_nsnp = 5000,
  plink = NULL,
  bfile = NULL,
  clump_r2 = 0.01,
  clump_kb = 10000,
  pop = "EUR",
  type = "exposure",
  cores = 1,
  cores_proxy = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{List of OpenGWAS IDs or file paths (to gwasvcf files)}

\item{rsids}{List of SNP rsIDs to extract}

\item{pval}{Threshold to extract SNPs (Optional)}

\item{proxies}{Whether to search for proxies (Optional, boolean)}

\item{proxy_rsq}{R2 threshold to use when searching for proxies (Optional)}

\item{proxy_kb}{kb threshold to use when searching for proxies (Optional)}

\item{proxy_nsnp}{Number of SNPs when searching for proxies (Optional)}

\item{plink}{Path to Plink binary (Optional)}

\item{bfile}{Path to Plink .bed/.bim/.fam files (Optional)}

\item{clump_r2}{r2 threshold for clumping SNPs (Optional)}

\item{clump_kb}{Distance outside of which SNPs are considered in linkage equilibrium (Optional)}

\item{pop}{Population (Optional, used only for clumping on OpenGWAS)}

\item{type}{Type of data (Optional, "exposure" or "outcome")}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{cores_proxy}{Number of cores for multi-threaded proxy searching (Optional)
NB: Unavailable on Windows machines
NB: Should not be more than `cores` argument!}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Data.frame of datasets
}
\description{
Helper function that is called from \link[mrpipeline]{read_exposure} and \link[mrpipeline]{read_outcome}.
Extracts exposure and outcome data according to arguments.
Should not be called directly.
}
\seealso{
[read_exposure()], [read_outcome()]
}
