% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{.wr_taylor_approx}
\alias{.wr_taylor_approx}
\title{Calculates the second term Taylor approximation for standard error of
the Wald ratio method.
From https://doi.org/10.1101/2021.03.01.433439 supplementary}
\usage{
.wr_taylor_approx(dat)
}
\arguments{
\item{object}{Data.frame or vector}
}
\value{
Appended result
}
\description{
Calculates the second term Taylor approximation for standard error of
the Wald ratio method.
From https://doi.org/10.1101/2021.03.01.433439 supplementary
}
