% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_target.r
\name{drug_target_evidence}
\alias{drug_target_evidence}
\title{Generate drug target evidence}
\usage{
drug_target_evidence(dat, ensg_col = "exposure")
}
\arguments{
\item{dat}{A data.frame or named list}

\item{ensg_col}{Column, or name, to be accessed in `dat`}
}
\value{
data.frame of results
}
\description{
Uses the Drug Genome Interaction DB's API to search for drug target-related
evidence, including on: Druggable Genome, Clinically Actionable and
Drug Resistant ontologies.
}
\details{
The lookup MUST be ENSG IDs.
}
