% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eaf_to_maf}
\alias{eaf_to_maf}
\title{Convert effect allele frequency to minor allele frequency}
\usage{
eaf_to_maf(eaf)
}
\arguments{
\item{eaf}{Numeric vector of effect allele frequencies.}
}
\value{
Numeric vector of minor allele frequencies.
}
\description{
Convert effect allele frequency to minor allele frequency
}
\keyword{internal}
