% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{experiment_heatmap}
\alias{experiment_heatmap}
\title{Create a heatmap plot}
\usage{
experiment_heatmap(df, palette = "OrRd", method_order, outcome_order,
  ...)
}
\arguments{
\item{df}{\code{data.frame} with following columns \itemize{
  \item{method}
  \item{outcome}
  \item{pval}
  \item{or}
  \item{ci_high}
  \item{ci_low}
}}

\item{palette}{\code{character} Color paletter to be passed to \code{pheatmap}}

\item{method_order}{\code{character} vector defining order of columns in the heatmap.
Should match values in the \code{method} column.}

\item{outcome_order}{\code{character} vector defining order of rows in the heatmap.
Should match values in the \code{outcome} column.}

\item{...}{Additional arguments passed to \code{\link[pheatmap]{pheatmap}}}
}
\value{
A plot returned by \code{\link[pheatmap]{pheatmap}}
}
\description{
Create a heatmap plot
}
