% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{extract_matched_regions}
\alias{extract_matched_regions}
\title{Extract SNPs based on region for colocalisation analyses.
Can be used before calling the `do_coloc` function or will be called as part
of that function automatically.}
\usage{
extract_matched_regions(dat, window = 5e+05, cores = 1, verbose = TRUE)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{window}{Window around SNPs to extract (Optional)}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Named list of matched regional data
}
\description{
Extract SNPs based on region for colocalisation analyses.
Can be used before calling the `do_coloc` function or will be called as part
of that function automatically.
}
\seealso{
[mrpipeline::do_coloc()]
}
