% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{f_plot}
\alias{f_plot}
\title{Plots F-statistic against P value
UNIMPLEMENTED}
\usage{
f_plot(
  dat,
  report,
  beta_col = "beta.exposure",
  se_col = "se.exposure",
  pval_col = "pval.exposure",
  trait_col = "exposure",
  id_col = "id.exposure"
)
}
\description{
Plots F-statistic against P value
UNIMPLEMENTED
}
