% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{file_to_gwasvcf}
\alias{file_to_gwasvcf}
\title{Convert file(s) to gwasvcf format.}
\usage{
file_to_gwasvcf(
  file,
  chr_col,
  pos_col,
  nea_col,
  ea_col,
  snp_col = NULL,
  eaf_col = NULL,
  beta_col = NULL,
  se_col = NULL,
  pval_col = NULL,
  n = NULL,
  n_case = NULL,
  name = NULL,
  header = TRUE,
  sep = "\\t",
  cores = 1,
  bcf_tools = NULL,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to file}

\item{chr_col}{Column name for chromosome}

\item{pos_col}{Column name for position}

\item{nea_col}{Column name for non-effect allele}

\item{ea_col}{Column name for effect allele}

\item{snp_col}{Column name for SNP (Optional)}

\item{eaf_col}{Column name for effect allele frequency (Optional)}

\item{beta_col}{Column name for beta (Optional)}

\item{se_col}{Column name for standard error (Optional)}

\item{pval_col}{Column name for P value (Optional)
NB: P values will be saved as 10^-P}

\item{n}{Sample size (Optional), can be int or column name}

\item{n_case}{Number of cases (Optional), can be int or column name}

\item{name}{Trait name (Optional), can be string or column name}

\item{header}{Whether file has a header or not (Optional, boolean)}

\item{sep}{File separater (Optional)}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{bcf_tools}{Path to bcf_tools (Optional)}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
gwasvcf object(s)
}
\description{
Function to convert a file (or files) to gwasvcf format.
}
\seealso{
[dat_to_gwasvcf()] for converting data.frames
}
