% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pqtl_decode.R
\name{format_decode_pqtl}
\alias{format_decode_pqtl}
\title{Formats deCODE genetics proteomic data for Mendelian Randomization analysis}
\usage{
format_decode_pqtl(
  decode_proteomic_gwas_file_path,
  decode_included_variants_file_path,
  pqtl_assay,
  x_y_chr_file = NULL
)
}
\arguments{
\item{decode_proteomic_gwas_file_path}{Dataframe, the file path to the file containing deCODE GWAS data}

\item{decode_included_variants_file_path}{Dataframe, the file path to the file containing the deCODE GWAS data for variants to include with suitable quality control}

\item{pqtl_assay}{String, the deCODE genetics protein assay}

\item{x_y_chr_file}{String, file path to the file containing rsids for X and Y chromosomes}
}
\value{
A list with two elements:
\itemize{
\item \code{exposure}: Formatted exposure data frame (output of TwoSampleMR::format_data).
}
}
\description{
This function filters deCODE genetics GWAS by the deCODE included variants file,
processes the data to clean the column headings and uses TwoSampleMR::format_data() to prepare
the exposure dataset before run_mr()
}
