% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_ukbppp.R
\name{format_mr_ukbppp}
\alias{format_mr_ukbppp}
\title{Formats data for using UKB-PPP data}
\usage{
format_mr_ukbppp(olink_linker_file, synapse_id, outcome, outcome_id, ref_rsid)
}
\description{
Formats data for using UKB-PPP data
}
