% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sc_RNA_eqtl.R
\name{format_sc_RNA_eqtl}
\alias{format_sc_RNA_eqtl}
\title{Formats single-cell RNA eQTL data frame}
\usage{
format_sc_RNA_eqtl(mapping_file, cell_type, eqtl_dir)
}
\description{
Purpose of this script of functions are to create the pathway to
the eQTL data and then format the eQTL data ready for analysis
Make sure list of cell_types are created/called from the mapping file
Read in eQTL data for individual cell type using mapping_file pathway
Rename variables and keep only ones of interest
Applying format_data()
}
