% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{get_col_name}
\alias{get_col_name}
\title{Get column names from agnostic but formatted data.frame}
\usage{
get_col_name(df, data)
}
\arguments{
\item{df}{Data.frame of formatted data (exposure or outcome)}

\item{data}{Name of column to find}
}
\value{
Name of column formatted for given data.frame
}
\description{
Get column names from agnostic but formatted data.frame
}
\keyword{Internal}
