% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{harmonise}
\alias{harmonise}
\title{Harmonise exposure and outcomes.}
\usage{
harmonise(exposure, outcome, action = 1, cores = 1)
}
\arguments{
\item{exposure}{Data.frame of exposure dataset(s)}

\item{outcome}{Data.frame of outcome dataset(s)}

\item{action}{How to harmonise alleles; see \link[TwoSampleMR]{harmonise_data}.}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}
}
\value{
Harmonised data.frame
}
\description{
Wrapper function for \link[TwoSampleMR]{harmonise_data} function in the
`TwoSampleMR` package.
}
\seealso{
[TwoSampleMR::harmonise_data()]
}
