% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{interactive_scatter_plot}
\alias{interactive_scatter_plot}
\title{Plots an interactive scatter plot of SNP-exposure and SNP-outcome effects}
\usage{
interactive_scatter_plot(dat, id.exposure, id.outcome)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{id.exposure}{Exposure ID}

\item{id.outcome}{Outcome ID}
}
\value{
Plot
}
\description{
Plots an interactive scatter plot of SNP-exposure and SNP-outcome effects
}
