% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{interactive_volcano_plot}
\alias{interactive_volcano_plot}
\title{Plots and returns an interactive volcano plot}
\usage{
interactive_volcano_plot(dat)
}
\arguments{
\item{dat}{A data.frame of harmonised data}
}
\value{
Plot
}
\description{
Plots and returns an interactive volcano plot
}
