% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{interactive_z_plot}
\alias{interactive_z_plot}
\title{Creates Z-score plot for SNPs}
\usage{
interactive_z_plot(dat)
}
\arguments{
\item{dat}{A data.frame of harmonised data}
}
\value{
Plot
}
\description{
Creates Z-score plot for SNPs
}
