% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{is_ambiguous}
\alias{is_ambiguous}
\title{Check if SNP frequencies are ambiguous}
\usage{
is_ambiguous(freq, tol = 0.08)
}
\arguments{
\item{freq}{Frequency}

\item{tol}{Tolerance around 0.5 (Optional)}
}
\value{
True/False if ambiguous
}
\description{
Check if SNP frequencies are ambiguous
}
