% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{is_palindromic}
\alias{is_palindromic}
\title{Check if SNP is palindromic}
\usage{
is_palindromic(A1, A2)
}
\arguments{
\item{A1}{Allele 1}

\item{A2}{Allele 2}
}
\value{
True/False if palindromic
}
\description{
Check if SNP is palindromic
}
