% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{make_outcomes}
\alias{make_outcomes}
\title{Make markdown report for each outcome}
\usage{
make_outcomes(filepath, filename, report, dat, id.outcome, trait.name)
}
\arguments{
\item{filepath}{Character, path to save}

\item{filename}{Character, name of file}

\item{report}{QCReport class of results, etc. for reporting}

\item{dat}{A data.frame of harmonised data}

\item{id.outcome}{ID of outcome}

\item{trait.name}{Annotated name of exposure}
}
\description{
Make markdown report for each outcome
}
