% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{make_report}
\alias{make_report}
\title{Make "main" markdown report}
\usage{
make_report(filepath, filename, dat, report)
}
\arguments{
\item{filepath}{Character, path to save}

\item{filename}{Character, name of file}

\item{dat}{A data.frame of harmonised data}

\item{report}{QCReport class of results, etc. for reporting}
}
\description{
Make "main" markdown report
}
