% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\name{make_results}
\alias{make_results}
\title{Make results markdown report for each exposure}
\usage{
make_results(filepath, filename, report, dat, id.exposure, trait.name)
}
\arguments{
\item{filepath}{Character, path to save}

\item{filename}{Character, name of file}

\item{report}{QCReport class of results, etc. for reporting}

\item{dat}{A data.frame of harmonised data}

\item{id.exposure}{ID of exposure}

\item{trait.name}{Annotated name of exposure}
}
\description{
Make results markdown report for each exposure
}
