% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{mr_scatter_plot}
\alias{mr_scatter_plot}
\title{Plots a scatter plot of MR results
UNIMPLEMENTED}
\usage{
mr_scatter_plot(res, keys, dat, report)
}
\description{
Plots a scatter plot of MR results
UNIMPLEMENTED
}
