% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpipeline-package.R
\docType{package}
\name{mrpipeline-package}
\alias{mrpipeline}
\alias{mrpipeline-package}
\title{mrpipeline: Functions to set up a pipeline for performing Mendelian randomization}
\description{
Pipeline to perform forward mendelian randomization analyses, including multiple comparison assessments and colocalization analyses. This package provides manuscript ready tables and figures.
}
\author{
\strong{Maintainer}: Benjamin Zuckerman \email{benjamin.zuckerman@kcl.ac.uk} (\href{https://orcid.org/0000-0002-0077-6074}{ORCID})

Authors:
\itemize{
  \item Alasdair Warwick (\href{https://orcid.org/0000-0002-0800-2890}{ORCID})
}

}
\keyword{internal}
