% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{pairwise_analysis}
\alias{pairwise_analysis}
\title{Performs pairwise harmonisation and analyses -- helpful when analysing many
exposure-outcome pairs as performing the standard "linear" approach will
be very slow.}
\usage{
pairwise_analysis(
  exposure,
  outcome,
  res_path,
  ...,
  do_coloc = FALSE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{exposure}{Data.frame of exposure dataset(s)}

\item{outcome}{Data.frame of outcome dataset(s)}

\item{res_path}{Path to save result files}

\item{...}{Other arguments for the following functions:
\link[mrpipeline]{harmonise}
\link[mrpipeline]{do_mr}
\link[mrpipeline]{do_coloc}}

\item{do_coloc}{True/False run colocalisation analyses}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\description{
In this function, an exposure-outcome pair are harmonised, analyses are ran
on those data and those results are saved to a file. Analyses ran can be MR
or colocalisation, as desired.
}
