% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{phewas_plot}
\alias{phewas_plot}
\title{Plots a PheWAS-like plot of all exposures against one outcome}
\usage{
phewas_plot(res, report)
}
\arguments{
\item{res}{A data.frame of results from MR}

\item{report}{QCReport class of results, etc. for reporting}
}
\description{
Plots a PheWAS-like plot of all exposures against one outcome
}
