% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coloc_pp_bar}
\alias{plot_coloc_pp_bar}
\title{Bar chart of ABF posterior probabilities}
\usage{
plot_coloc_pp_bar(x)
}
\arguments{
\item{x}{A \code{coloc_result} object.}
}
\value{
A ggplot object or \code{NULL} invisibly.
}
\description{
Bar chart of ABF posterior probabilities
}
\keyword{internal}
