% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coloc_regional}
\alias{plot_coloc_regional}
\title{Regional association plot for coloc results}
\usage{
plot_coloc_regional(x)
}
\arguments{
\item{x}{A \code{coloc_result} object.}
}
\value{
A ggplot object or \code{NULL} invisibly.
}
\description{
Regional association plot for coloc results
}
\keyword{internal}
