% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot pannel of scatter plots}
\usage{
plot_scatter(result, mar)
}
\arguments{
\item{result}{a \code{list} as returned from \code{\link{compute_result_for_exposure}}}

\item{mar}{a numeric vector suitable as an argument for \code{par(mar = mar)}}
}
\value{
A set of plots
}
\description{
The result contains four plots \itemize{
  \item{Scatterplot showing causal effect estimates}
  \item{Funnel plot showing individual SNP level causal effect estimates}
  \item{Heterogeneity statistic of individual SNPs}
  \item{Radial MR plot showing potential outlier SNPs}
}
}
