% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{q_test}
\alias{q_test}
\title{Cochran Q-test (p-value) of instrument i.e. trait}
\usage{
q_test(y, s)
}
\arguments{
\item{y}{\code{numeric}}

\item{s}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Cochran Q-test (p-value) of instrument i.e. trait
}
