% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\docType{class}
\name{QCReport-class}
\alias{QCReport-class}
\alias{QCReport}
\title{A Reference Class for the reporting of QC, results, etc.}
\description{
A Reference Class for the reporting of QC, results, etc.
}
\section{Fields}{

\describe{
\item{\code{exposures}}{A data.frame of QC results for the exposure datasets}

\item{\code{outcomes}}{A data.frame of QC results for the outcome datasets}

\item{\code{results}}{A data.frame of MR results}

\item{\code{sresults}}{A data.frame of MR sensitivity results}

\item{\code{hetresults}}{A data.frame of heterogeneity results}

\item{\code{pleioresults}}{A data.frame of pleiotropy results}

\item{\code{steigerresults}}{A data.frame of Steiger filtering results}

\item{\code{cresults}}{A data.frame of colocalisation results}

\item{\code{bonferroni}}{Numeric, Bonferroni-corrected P value threshold}

\item{\code{plots}}{A data.frame of plot names, IDs, etc.}

\item{\code{raw.plots}}{A list of all plots}
}}

\section{Methods}{

\describe{
\item{\code{add_cresults(r)}}{Adds colocalisation reuslts to data.frame}

\item{\code{add_dataset(name, x)}}{Add dataset to either exposure/outcome data.frame}

\item{\code{add_hresults(r)}}{Adds heterogeneity results to data.frame}

\item{\code{add_plot(p, raw)}}{Adds information and raw plot to report}

\item{\code{add_results(r)}}{Adds MR results to data.frame}

\item{\code{add_sresults(name, x)}}{Adds sensitivity MR results to data.frame}

\item{\code{get_exposures_name()}}{Internal function to get name of exposure data.frame}

\item{\code{get_outcomes_name()}}{Internal function to get name of outcome data.frame}

\item{\code{initialise()}}{Initialise all fields}

\item{\code{update_cell(name, trait, col, val)}}{Update cell of exposure/outcome data.frame}
}}

