% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{read_datasets}
\alias{read_datasets}
\title{Reads exposure and outcome datasets from local .vcf files
or OpenGWAS DB}
\usage{
read_datasets(
  id1,
  id2,
  report,
  p_cutoff,
  chrompos_query,
  gene_query,
  f_cutoff = 10,
  r2 = 0.001,
  kb = 10000,
  pop = "EUR",
  proxies = T,
  action = 1,
  bfile = NULL,
  plink_bin = NULL,
  nthreads = 1
)
}
\arguments{
\item{id1}{DatasetsID class of exposure IDs}

\item{id2}{DatasetsID class of outcome IDs}

\item{report}{QCReport class of results, etc. for reporting}

\item{p_cutoff}{P value cutoff for "tophits" from summary statistics}

\item{chrompos_query}{Select SNPs within this region only,
format: "chrom:start-end"}

\item{gene_query}{Select SNPs nearby this gene only, NOT YET IMPLEMENTED}

\item{f_cutoff}{F-statistic cutoff for exposure data, default: 10}

\item{r2}{Clumping threshold, default: 0.001}

\item{kb}{Clumping window, default: 10000}

\item{pop}{Super-population for clumping, default: EUR}

\item{proxies}{Whether to search for and use proxy SNPs in the outcome
dataset, default: TRUE}

\item{action}{See: TwoSampleMR::harmonise_data, default: 1}

\item{bfile}{Plink 1.x format reference files for use in local clumping,
etc, default: NULL}

\item{plink_bin}{Plink 1.x location, default: NULL}

\item{nthreads}{Number of threads for use in multithreaded functions,
default: 1}
}
\value{
Data.frame of harmonised exposure-outcome SNPs
}
\description{
Reads exposure and outcome datasets from local .vcf files
or OpenGWAS DB
}
