% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{read_exposure}
\alias{read_exposure}
\title{Read exposures}
\usage{
read_exposure(
  ids,
  pval = 5e-08,
  plink = NULL,
  bfile = NULL,
  clump_r2 = 0.01,
  clump_kb = 10000,
  pop = "EUR",
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{List of OpenGWAS IDs or file paths (to gwasvcf files)}

\item{pval}{Threshold to extract SNPs (Optional)}

\item{plink}{Path to Plink binary (Optional)}

\item{bfile}{Path to Plink .bed/.bim/.fam files (Optional)}

\item{clump_r2}{r2 threshold for clumping SNPs (Optional)}

\item{clump_kb}{Distance outside of which SNPs are considered in linkage equilibrium (Optional)}

\item{pop}{Population (Optional, used only for clumping on OpenGWAS)}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Data.frame of exposure datasets
}
\description{
Read a dataset (or datasets) as exposure.
Can accept both OpenGWAS IDs or file paths. Accepts only gwasvcf file formats.
This function can clump data locally, if supplied with the `plink` and `bfile`
arguments. If these are not supplied, clumping will take place on the
OpenGWAS servers \emph{only} for OpenGWAS IDs.
If you would like to clump local files, please provide paths to Plink and bfiles.
}
