% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.r
\name{read_outcome}
\alias{read_outcome}
\title{Read outcomes}
\usage{
read_outcome(
  ids,
  rsids,
  proxies = TRUE,
  proxy_rsq = 0.8,
  proxy_kb = 5000,
  proxy_nsnp = 5000,
  plink = NULL,
  bfile = NULL,
  cores = 1,
  cores_proxy = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{List of OpenGWAS IDs or file paths (to gwasvcf files)}

\item{rsids}{List of SNP rsIDs to extract}

\item{proxies}{Whether to search for proxies (Optional, boolean)}

\item{proxy_rsq}{R2 threshold to use when searching for proxies (Optional)}

\item{proxy_kb}{kb threshold to use when searching for proxies (Optional)}

\item{proxy_nsnp}{Number of SNPs when searching for proxies (Optional)}

\item{plink}{Path to Plink binary (Optional)}

\item{bfile}{Path to Plink .bed/.bim/.fam files (Optional)}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{cores_proxy}{Number of cores for multi-threaded proxy searching (Optional)
NB: Unavailable on Windows machines
NB: Should not be more than `cores` argument!}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
Data.frame of outcome datasets
}
\description{
Read a dataset (or datasets) as outcome.
Can accept both OpenGWAS IDs or file paths. Accepts only gwasvcf file formats.
This function can search for proxy SNPs locally, if supplied with the `plink`
and `bfile` arguments. If these are not supplied, proxy searching will take
place on the OpenGWAS servers \emph{only} for OpenGWAS IDs.
If you would like to search for proxies locally, please provide paths to Plink
and bfiles.
}
