% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{regional_plot}
\alias{regional_plot}
\title{Plots a regional plot of the area being tested for colocalisation}
\usage{
regional_plot(
  dat,
  exposure,
  outcome,
  bfile = NULL,
  plink = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{exposure}{Character, name of exposure}

\item{outcome}{Character, name of outcome}

\item{bfile}{Path to Plink bed/bim/fam files}

\item{plink}{Path to Plink binary}

\item{verbose}{Print messages or not}
}
\description{
Plots a regional plot of the area being tested for colocalisation
}
