% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{report_mr}
\alias{report_mr}
\title{Read config file and knit the report}
\usage{
report_mr(config, ...)
}
\arguments{
\item{config}{character path to config.yaml.}

\item{...}{aditional arguments passed to \code{\link[rmarkdown]{render}}}
}
\value{
see \code{\link[rmarkdown]{render}}
}
\description{
The \code{config} file should have a following structure
\preformatted{
rnd_seed: 123
input_path: "/path/to/input.xlsx"
metadata:
  n_out: 66164
  n_cas: 4127
  n_exp:
    Trait01: 518633
    Trait02: 370711
    Trait03: 318633
    Trait04: 290711
    Trait05: 420711
}
}
