% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{standardize_names}
\alias{standardize_names}
\title{Convert column names of a data.frame to a query friendly format}
\usage{
standardize_names(df)
}
\arguments{
\item{df}{data.frame Frame to be processed}
}
\value{
data.frame with query friendly names
}
\description{
This function converts all names to lower case, remove ascii characters, replaces spaces with underscores.
}
\examples{
standardize_names(tibble::tibble(
  `Proxy   SNP id_$` = "rs1234",
  `Proxy  SNP id`= "rs1234",
  `Proxy__SNP_id_` = "rs1234",
  `Proxy_SNP_id_` = "rs1234"))
}
