% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{standarize_names}
\alias{standarize_names}
\title{Convert column names of a data.frame to a query friendly format}
\usage{
standarize_names(df)
}
\description{
Convert column names of a data.frame to a query friendly format
}
