% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utilis.R
\name{tangram_results}
\alias{tangram_results}
\title{Conver MR analysis results to tangram object}
\usage{
tangram_results(results, outcome)
}
\arguments{
\item{results}{list of the same shape as returned from \code{\link{run_mr}}}

\item{outcome}{character}
}
\value{
a tangram object (see \code{\link[tangram]{tangram}} for details)
}
\description{
Conver MR analysis results to tangram object
}
