% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_pipeline.r
\name{validate_config}
\alias{validate_config}
\title{Validates parameters in config file}
\usage{
validate_config(conf)
}
\arguments{
\item{conf}{config::config file of parameters}
}
\value{
Validated config class
}
\description{
Validates parameters in config file
}
