% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mr.R
\name{validate_exclude_regions}
\alias{validate_exclude_regions}
\title{Validate exclude_regions argument}
\usage{
validate_exclude_regions(exclude_regions)
}
\arguments{
\item{exclude_regions}{Data frame to validate.}
}
\value{
Invisibly returns \code{TRUE} if valid; otherwise aborts with an error.
}
\description{
Validate exclude_regions argument
}
\keyword{internal}
