% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{weighted.median.boot}
\alias{weighted.median.boot}
\title{Compute MC esitmate of weighted median}
\usage{
weighted.median.boot(betaXG.in, betaYG.in, sebetaXG.in, sebetaYG.in,
  weights.in)
}
\arguments{
\item{betaXG.in}{\code{numeric}}

\item{betaYG.in}{\code{numeric}}

\item{sebetaXG.in}{\code{numeric}}

\item{sebetaYG.in}{\code{numeric}}

\item{weights.in}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Used by compute_weighted_median with
\itemize{
  \item{\code{effect_of_lead_variant_on_exposure_level} as \code{betaXG.in}}
  \item{\code{effect_of_lead_variant_on_outcome_levels} as \code{betaYG.in}}
  \item{\code{standard_error_of_effect_on_exposure_se} as \code{sebetaXG}}
  \item{\code{standard_error_of_effect_on_outcome_se} as \code{sebetaXG.in}}
  \item{\code{sqrt(standard_error_of_effect_on_outcome_se / effect_of_lead0variant_on_exposure_level)}}
}
}
