% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{z_plot}
\alias{z_plot}
\title{Z plot}
\usage{
z_plot(
  dat,
  snp_col = "SNP",
  beta_col = "beta.exposure",
  se_col = "se.exposure",
  pval_col = "pval.exposure",
  force_static = FALSE
)
}
\arguments{
\item{dat}{A data.frame of data}

\item{snp_col}{Column of SNP names (Optional)}

\item{beta_col}{Column of MR beta estimates (Optional)}

\item{se_col}{Column of standard errors for the beta estimates (Optional)}

\item{pval_col}{Column of P values (Optional)}

\item{force_static}{True for forcing the plot to be returned as a static plot (Optional)}
}
\value{
Plot
}
\description{
Creates a Z-score plot for SNPs, where \eqn{Z = b / SE}.
This should follow a parabollic shape and so can be used to find certain
SNPs which may not follow this shape.
If `plotly` is installed, the plot will be returned interactive.
}
