% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{combine_experiment_data}
\alias{combine_experiment_data}
\title{Given single outcome results, return data.frame suitable for experiment_heatmap}
\usage{
combine_experiment_data(outcome, mr, heterogeneity_metrics, meta,
  mappers = list(egger = function(x) x["b1", ]))
}
\description{
Given single outcome results, return data.frame suitable for experiment_heatmap
}
