% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_utils.R
\name{combine_experiments_data}
\alias{combine_experiments_data}
\title{Combines multiple exeperiments data}
\usage{
combine_experiments_data(results)
}
\description{
Combines multiple exeperiments data
}
