% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{compute_mbe}
\alias{compute_mbe}
\title{Weighted median estimate}
\usage{
compute_mbe(BetaXG, BetaYG, seBetaXG, seBetaYG, phi = c(1),
  n_boot = 10000, alpha = 0.05)
}
\description{
Weighted median estimate
}
