% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{do_coloc}
\alias{do_coloc}
\title{Run colocalisation analyses}
\usage{
do_coloc(
  dat,
  method = "coloc.abf",
  coloc_window = 5e+05,
  bfile = NULL,
  plink = NULL,
  pwcoco = NULL,
  workdir = tempdir(),
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{method}{Which method of colocalisation to use: coloc.abf, coloc.susie, pwcoco (Optional)}

\item{coloc_window}{Size (+/-) of region to extract for colocalisation analyses (Optional)}

\item{bfile}{Path to Plink bed/bim/fam files (Optional)}

\item{plink}{Path to Plink binary (Optional)}

\item{pwcoco}{If PWCoCo is the selected coloc method, path to PWCoCo binary (Optional)}

\item{workdir}{Path to save temporary files (Optional)}

\item{cores}{Number of cores for multi-threaded tasks (Optional)
NB: Unavailable on Windows machines}

\item{verbose}{Display verbose information (Optional, boolean)}
}
\value{
A data.frame of colocalistion results
}
\description{
Runs colocalisation using any of the following methods:
\enumerate{
\item Coloc.abf, see coloc::coloc.abf()
\item Coloc.susie, see coloc::coloc.susie()
\item PWCoCo, see \href{https://github.com/jwr-git/pwcoco}{PWCoCo}
}
NB: PWCoCo is not available on Windows.
}
\seealso{
[coloc::coloc.abf()], [coloc::coloc.susie()]
}
