% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.r
\name{.wr_taylor_approx}
\alias{.wr_taylor_approx}
\title{WR Taylor Approx of SE}
\usage{
.wr_taylor_approx(dat)
}
\arguments{
\item{object}{Harmonised data.frame}
}
\value{
Results data.frame
}
\description{
Calculates the second term Taylor approximation for standard error of
the Wald ratio method.
From \href{https://doi.org/10.1101/2021.03.01.433439}{supplementary}
}
\keyword{Internal}
