% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pqtl_decode.R
\name{format_pqtl_decode}
\alias{format_pqtl_decode}
\title{Formats deCODE genetics proteomic data for Mendelian Randomization analysis}
\usage{
format_pqtl_decode(
  decode_proteomic_gwas_file_path,
  decode_included_variants_file_path,
  pqtl_assay,
  x_y_chr_file = NULL
)
}
\arguments{
\item{decode_proteomic_gwas_file_path}{Character vector of file path(s) to deCODE GWAS data,
or a single pre-loaded data.frame. If multiple paths are provided, data will be combined.}

\item{decode_included_variants_file_path}{Character vector of file path(s) to the deCODE
included variants data, or a single pre-loaded data.frame. If multiple paths are provided,
data will be combined.}

\item{pqtl_assay}{String, name of the deCODE genetics protein assayed}

\item{x_y_chr_file}{String, optional file path to a file containing rsids for X and Y
chromosomes. This file should be tab-separated with columns: Chromosome, Position, RSID.}
}
\value{
A list with two elements:
\itemize{
\item \code{exposure}: Formatted exposure data frame (output of TwoSampleMR::format_data).
}
}
\description{
This function filters deCODE genetics GWAS by the deCODE included variants file,
processes the data to clean the column headings and uses TwoSampleMR::format_data() to prepare
the exposure dataset before run_mr()
}
\examples{
\dontrun{
result <- format_pqtl_decode(
  decode_proteomic_gwas_file_path = "path/to/decode_gwas.txt.gz",
  decode_included_variants_file_path = "path/to/included_variants.txt.gz",
  pqtl_assay = "CD40"
)
head(result$exposure)
}

}
