% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ukbppp_file.R
\name{format_pqtl_ukbppp}
\alias{format_pqtl_ukbppp}
\title{Formats pQTL data from UKB-PPP for MR analysis.}
\usage{
format_pqtl_ukbppp(ukbppp, ukbppp_rsid, pqtl_assay)
}
\arguments{
\item{ukbppp}{Dataframe, the ukbppp data}

\item{ukbppp_rsid}{Dataframe, of ukbppp rsids}

\item{pqtl_assay}{String, of the ukbppp protein assayed}
}
\value{
A list with two elements:
\itemize{
\item \code{exposure}: Formatted exposure data frame (output of TwoSampleMR::format_data).
}
}
\description{
This function processes pQTL summary statistics from UKB-PPP,
matches them with rsID information, handles non-Mendelian
chromosomes, standardizes column names, and formats the data
for use with the TwoSampleMR package.
}
\examples{
# See the test script for example usage.
TO DO
}
