% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_single_cell_onek1k.R
\name{format_single_cell_onek1k}
\alias{format_single_cell_onek1k}
\title{Format single-cell RNA eQTL data for analysis}
\usage{
format_single_cell_onek1k(mapping_file, cell_type, eqtl_dir)
}
\arguments{
\item{mapping_file}{Character. File path to the mapping file containing eQTL data paths.}

\item{cell_type}{Character. Name of the cell type to subset the eQTL data.}

\item{eqtl_dir}{Character or data.frame. Either the path to the eQTL data directory or a pre-loaded data frame.}
}
\value{
A list with one data frame containing the formatted single cell RNA data
}
\description{
This function reads in eQTL data for a specified cell type using a provided mapping file
and formats the data frame to prepare it for downstream analysis.
The function expects the eQTL data directory path and uses \code{TwoSampleMR::format_data}
to structure the data in a standardized format.
}
\examples{
# See the test script for example usage.
Read in eQTL data for individual cell type using mapping_file pathway
Rename variables and keep only ones of interest
Applying format_data()
}
