% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_single_cell_onek1k.R
\name{format_single_cell_onek1k}
\alias{format_single_cell_onek1k}
\title{Formats single-cell RNA eQTL data frame}
\usage{
format_single_cell_onek1k(mapping_file, cell_type, eqtl_dir)
}
\description{
Purpose of this script of functions are to create the pathway to
the eQTL data and then format the eQTL data ready for analysis
Make sure list of cell_types are created/called from the mapping file
Read in eQTL data for individual cell type using mapping_file pathway
Rename variables and keep only ones of interest
Applying format_data()
}
