% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_pipeline.r
\name{mr_pipeline}
\alias{mr_pipeline}
\title{Entry point for the pipeline.}
\usage{
mr_pipeline(ids1, ids2, out_name = "", config_file = "", ...)
}
\arguments{
\item{ids1}{IDs or filenames for summary statistics}

\item{ids2}{IDs or filenames for summary statistics}

\item{out_name}{Name of the analysis given to the markdown report,
defaults to time and date}

\item{config_file}{Path to config.yml file. Defaults to file that comes with the
package. Please see that file for more details.}

\item{...}{Other arguments for plotting, NOT YET IMPLEMENTED}
}
\value{
list of results for debugging
}
\description{
Entry point for the pipeline.
}
