% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{regional_plot}
\alias{regional_plot}
\title{Plots a regional plot of the area being tested for colocalisation}
\usage{
regional_plot(dat, pairs, i, report)
}
\arguments{
\item{dat}{A data.frame of harmonised data}

\item{pairs}{A data.frame of pairwise combined IDs}

\item{i}{Location in pairs being tested}

\item{report}{QCReport class of results, etc. for reporting}
}
\description{
Plots a regional plot of the area being tested for colocalisation
}
