% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mr.R
\name{run_mr}
\alias{run_mr}
\title{Performs MR}
\usage{
run_mr(
  exposure,
  exposure_id,
  outcome,
  outcome_id,
  mhc_remove,
  instrument_region = list(chromosome = 1L, start = 1L, end = 200L),
  window = 1L,
  pval_thresh = 5e-06,
  rsq_thresh = 0.1,
  bfile
)
}
\arguments{
\item{exposure}{data frame. Summary statistics for exposure, must include exposure_id column ie gene name which we use as above?}

\item{exposure_id}{character. e.g. "syn52361761" -> we should change this to the protein/gene name}

\item{outcome}{data frame. Summary statistics for outcome, must be formatted as above}

\item{outcome_id}{character. Name for outcome e.g. 'SjD'}

\item{mhc_remove}{Logical, whether to remove the MHC region}

\item{instrument_region}{list of the chromosome position, gene start and gene end for each gene of interest this needs to link with the original mapping file}

\item{window}{integer. Set to cis region if required}

\item{pval_thresh}{number. 5e-6 by default}

\item{rsq_thresh}{R square clumping threshold}

\item{bfile}{path to the LD_folder containing 1000 genome files (g1000_eur)}
}
\value{
List with 2 data frames - MR results and instruments
}
\description{
Performs MR
}
\examples{
To filter out MHC region
To filer for appropriate gene window
This needs to be done before the function starts
Will need to make sure this is on the format_data() outcome data ie chr.outcome and chr.exposure
Make sure no NA present in the exposure or outcome SNPs beta or se columns
Filtering mr_keep for TRUE
}
