% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{run_mr}
\alias{run_mr}
\title{Main entry point for MR pipeline}
\usage{
run_mr(df, meta)
}
\arguments{
\item{df}{data.frame A \code{data.frame} containing at least following columns
\itemize{
 \item{trait} {\code{character}}
 \item{effect_of_lead_variant_on_exposure_levels} {\code{numeric}}
 \item{effect_of_lead_variant_on_outcome_levels} {\code{numeric}}
 \item{standard_error_of_effect_on_exposure_se} {\code{numeric}}
 \item{standard_error_of_effect_on_outcome_se} {\code{numeric}}
}}

\item{meta}{list A \code{list} with at least following
\itemize{
 \item{n_out} {\code{numeric}} of length equal to 1
 \item{n_cas} {\code{numeric}} of length equal to 1
 \item{n_exp} {A named \code{list}} of {\code{numeric}} values
}}
}
\value{
A \code{mrresults} objec (named {\code{list}} which maps a trait to MR result).
}
\description{
Main entry point for MR pipeline
}
